+++ tools/sdk-ai-bots/azure-sdk-qa-bot-function/scripts/add-conversation-record-to-azure-table.ts
@@ -1,7 +1,7 @@
 import { v4 as uuidv4 } from "uuid";
 import { TableEntity } from "@azure/data-tables";
 import { Converter } from "./utils";
-import { loadChannelMapping } from "../src/common/shared";
+import { loadChannelMapping } from "../src/common/channelConfig";
 import { BlobService } from "../src/services/StorageService";
 
 interface RecordRow {
