+++ tools/sdk-ai-bots/azure-sdk-qa-bot-function/src/common/channelConfig.ts
